+++ src/shared/disk-encrypt.c
@@ -175,7 +175,7 @@ int disk_encrypt_get_info(FILE *f,
         return 0;
 }
 
-static int dm_ioctl_new(uint64_t device, unsigned int flags, size_t data_size, struct dm_ioctl **iop) {
+static int dm_ioctl_new(uint64_t device, unsigned int flags, uint64_t data_size, struct dm_ioctl **iop) {
         struct dm_ioctl *io;
 
         io = calloc(1, sizeof(struct dm_ioctl) + data_size);
